+++ src/lib/panel-secondary/tabs/settings.svelte
@@ -16,7 +16,7 @@
 	const {
 		elements: { content, item, trigger, root },
 		helpers: { isSelected }
-	} = createAccordion()
+	} = createAccordion({})
 
 	const items = [
 		{